feat: 게시글 상세 페이지 주소 변경

+++ jsp_board-2508/src/main/java/com/ll/jsp/board/boundedContext/article/repository/ArticleRepository.java
@@ -38,7 +38,7 @@ public class ArticleRepository {
     }
 
 
-    public Article findById(int id) {
+    public Article findById(Long id) {
         return articleList.stream()
                 .filter(article -> article.getId() == id)
                 .findFirst()
